agent_evaluation/requirements.txt
# Requirements for LangChain version

# Original dependencies
langfuse
litellm
python-dotenv
boto3
pydantic>=2.0.0

# LangChain core dependencies (v1.x+ uses langgraph for agents)
langchain>=1.0.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langgraph>=0.1.0

# LangChain model integrations
langchain-openai>=0.0.5
langchain-aws>=0.1.0  # for Bedrock

# Optional: Strands (if you want to keep compatibility)
# strands-agents
# strands-agents[openai]
# strands-agents[litellm]
# strands-agents-tools
